Create Down to Zero ||
This problem is an optimization puzzle focused on finding the shortest path from a number N down to 0. You are given two types of moves:

Decrement: A safe, single-step move where you change N to N-1.

Factorization Jump: A potential shortcut where, if N can be factored into a * b, you can instantly change N to the larger of the two factors.

The challenge is to determine the optimal sequence of these moves to reach 0 in the minimum possible steps. This can be modeled as a shortest path problem on a graph, making it well-suited for a Breadth-First Search (BFS) approach.