Create Die Hard 3
The solution to the water jug problem is based on a mathematical theorem known as Bézout's identity. It states that the smallest positive amount you can measure is the greatest common divisor (GCD) of the two jug capacities, a and b. Consequently, any amount c that can be measured must be a multiple of gcd(a, b).

Additionally, the target amount c cannot be larger than the bigger of the two jugs, since you must be able to hold the final amount in one of them.

Therefore, a solution is "YES" if two conditions are met:

c is less than or equal to the maximum of a and b.

c is perfectly divisible by the gcd(a, b).

Otherwise, the answer is "NO".