Anagram
The **Anagram** problem requires you to analyze a string to determine the minimum number of character changes needed to make its two equal-length halves anagrams of each other. The first step is crucial: if the string's length is **odd**, a split into two equal, contiguous substrings is impossible, so you must return **-1**. If the length is even, you split the string into $S_1$ and $S_2$. The solution then hinges on **frequency counting**: you count the occurrences of every character ('a' through 'z') in both $S_1$ and $S_2$. The minimum number of changes is calculated by summing the **excess** characters in $S_1$ that are not present in $S_2$. Specifically, for each character $c$, if the count in $S_1$ is greater than the count in $S_2$, the difference ($count_{S1}[c] - count_{S2}[c]$) represents the number of characters in $S_1$ that must be replaced to achieve the matching character set of an anagram.